Move VerifiablePresentation into separate file with its helpers + made tests for helpers + fixed bug
